+++ Sources/Views/Cells/MessageCollectionViewCell.swift
@@ -57,7 +57,6 @@ open class MessageCollectionViewCell: UICollectionViewCell, CollectionViewReusab
         super.init(frame: frame)
         contentView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
         setupSubviews()
-        setupGestureRecognizers()
     }
 
     required public init?(coder aDecoder: NSCoder) {
@@ -90,25 +89,33 @@ open class MessageCollectionViewCell: UICollectionViewCell, CollectionViewReusab
         }
     }
 
-    open func configureAvatar(_ avatar: Avatar) {
-        avatarView.set(avatar: avatar)
-    }
+    open func configure(with message: MessageType, at indexPath: IndexPath, and messagesCollectionView: MessagesCollectionView) {
+        guard let dataSource = messagesCollectionView.messagesDataSource else {
+            fatalError("MessagesDataSource is not set.")
+        }
+        guard let displayDelegate = messagesCollectionView.messagesDisplayDelegate else {
+            fatalError("MessagesDisplayDelegate is not set.")
+        }
 
-    open func configureAccessoryLabels(_ topText: NSAttributedString?, _ bottomText: NSAttributedString?) {
+        delegate = messagesCollectionView.messageCellDelegate
+
+        let messageColor = displayDelegate.backgroundColor(for: message, at: indexPath, in: messagesCollectionView)
+        let messageStyle = displayDelegate.messageStyle(for: message, at: indexPath, in: messagesCollectionView)
+        
+        displayDelegate.configureAvatarView(avatarView, for: message, at: indexPath, in: messagesCollectionView)
+
+        messageContainerView.backgroundColor = messageColor
+        messageContainerView.style = messageStyle
+
+        let topText = dataSource.cellTopLabelAttributedText(for: message, at: indexPath)
+        let bottomText = dataSource.cellBottomLabelAttributedText(for: message, at: indexPath)
+
         cellTopLabel.attributedText = topText
         cellBottomLabel.attributedText = bottomText
     }
 
-    func setupGestureRecognizers() {
-        let tapGesture = UITapGestureRecognizer(target: self, action: #selector(handleTapGesture(_:)))
-        contentView.addGestureRecognizer(tapGesture)
-    }
-
     /// Handle tap gesture on contentView and its subviews like messageContainerView, cellTopLabel, cellBottomLabel, avatarView ....
-    @objc
     open func handleTapGesture(_ gesture: UIGestureRecognizer) {
-        guard gesture.state == .ended else { return }
-
         let touchLocation = gesture.location(in: self)
 
         switch true {